style: :art: moved formatBerlinTime in utils file
