+++ app/frontend/src/features/message/utils.ts
@@ -1,21 +1,10 @@
 import { createId } from "@application-project-ws24/cuid";
-import { Temporal } from "@js-temporal/polyfill";
 import type { Message } from "@server/db/messages";
+import { formatBerlinTime } from "@server/lib/utils";
 
 // * extends server message with receivedAt
 export interface LocalMessage extends Message {
 	receivedAt: string;
-}
-
-export function formatBerlinTime() {
-	return Temporal.Now.zonedDateTimeISO("Europe/Berlin").toLocaleString(
-		"de-DE",
-		{
-			hour: "2-digit",
-			minute: "2-digit",
-			hour12: false,
-		},
-	);
 }
 
 export function createMessage(
